API - slot, location, products

+++ controllers/Location.js
@@ -1,12 +1,40 @@
 const Location = require("../models/Location");
+const Barber = require("../models/Barber");
+
+const { tokenValidation } = require("../tools");
 
 // Create a new location
 const createLocation = async (req, res) => {
     try {
-        const location = new Location(req.body);
+        const token = req.headers.authorization;
+        const decodedToken = await tokenValidation(token);
+
+        if (!decodedToken) {
+            res.status(401).json({ message: "Unauthorized" });
+            return;
+        }
+        const { email } = decodedToken;
+
+        let barber = await Barber.findOne({ email });
+        if (!barber) {
+            return res.status(401).json({ message: "Unauthorized" });
+        }
+
+        const { name, address, city, country, coordinates } = req.body
+
+        const location = new Location({
+            name,
+            address,
+            city,
+            country,
+            coordinates,
+            barber: barber._id
+        });
+
         await location.save();
         res.status(201).send(location);
     } catch (err) {
+        console.log(err)
         res.status(400).send(err);
     }
 };
@@ -14,17 +42,21 @@ const createLocation = async (req, res) => {
 // Get all locations
 const getAllLocations = async (req, res) => {
     try {
-        const locations = await Location.find();
-        res.send(locations);
-    } catch (err) {
-        res.status(500).send(err);
-    }
-};
+        const token = req.headers.authorization;
+        const decodedToken = await tokenValidation(token);
 
-// Get a single location by ID
-const getLocationById = async (req, res) => {
-    try {
-        const location = await Location.findById(req.params.id);
+        if (!decodedToken) {
+            res.status(401).json({ message: "Unauthorized" });
+            return;
+        }
+        const { email } = decodedToken;
+
+        let barber = await Barber.findOne({ email });
+        if (!barber) {
+            return res.status(401).json({ message: "Unauthorized" });
+        }
+
+        const location = await Location.find({ barber: barber._id });
         if (!location) {
             res.status(404).send("Location not found");
         } else {
@@ -38,9 +70,31 @@ const getLocationById = async (req, res) => {
 // Update a location by ID
 const updateLocation = async (req, res) => {
     try {
+        const token = req.headers.authorization;
+        const decodedToken = await tokenValidation(token);
+
+        if (!decodedToken) {
+            res.status(401).json({ message: "Unauthorized" });
+            return;
+        }
+        const { email } = decodedToken;
+
+        let barber = await Barber.findOne({ email });
+        if (!barber) {
+            return res.status(401).json({ message: "Unauthorized" });
+        }
+
+        const { name, address, city, country, coordinates } = req.body
+
         const location = await Location.findByIdAndUpdate(
             req.params.id,
-            req.body,
+            {
+                name,
+                address,
+                city,
+                country,
+                coordinates
+            },
             {
                 new: true,
                 runValidators: true,
@@ -52,6 +106,7 @@ const updateLocation = async (req, res) => {
             res.send(location);
         }
     } catch (err) {
+        console.log(err);
         res.status(400).send(err);
     }
 };
@@ -59,6 +114,20 @@ const updateLocation = async (req, res) => {
 // Delete a location by ID
 const deleteLocation = async (req, res) => {
     try {
+        const token = req.headers.authorization;
+        const decodedToken = await tokenValidation(token);
+
+        if (!decodedToken) {
+            res.status(401).json({ message: "Unauthorized" });
+            return;
+        }
+        const { email } = decodedToken;
+
+        let barber = await Barber.findOne({ email });
+        if (!barber) {
+            return res.status(401).json({ message: "Unauthorized" });
+        }
+
         const location = await Location.findByIdAndDelete(req.params.id);
         if (!location) {
             res.status(404).send("Location not found");
@@ -87,4 +156,4 @@ const getBarberLocations = async (req, res) => {
 };
 
 
-module.exports = { createLocation, getAllLocations, getLocationById, updateLocation, deleteLocation, getBarberLocations };
+module.exports = { createLocation, getAllLocations, updateLocation, deleteLocation, getBarberLocations };
